do sth

+++ delete_account/delete_account.py
@@ -174,6 +174,8 @@ def switch():
 
 if __name__ == '__main__':
     logging.basicConfig(level=logging.INFO)
+    if not os.path.exists('ssrmu.sh'):
+        raise Exception('请将ssrmu.sh脚本与本脚本放于同一目录中')
     if len(sys.argv) == 1:
         delete_acc_by_file()
         sys.exit(0)
